requirements.txt
groq
python-dotenv
tenacity
pytest
pytest-asyncio